Refactoring: Introduce new color module.
